% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustergraph_rnd.R
\name{clustergraph_rnd}
\alias{clustergraph_rnd}
\title{Computer cluster test based on multiple signal.}
\usage{
clustergraph_rnd(formula, data, signal, method, threshold, np, P, graph,
  effect = NULL, coding_sum = T, aggr_FUN, multcomp = "maris_oostenveld",
  return_distribution, ...)
}
\arguments{
\item{formula}{the formula of the model}

\item{data}{a dataframe containing the design}

\item{signal}{a 3 dimentional array. The the row are the observations, colomn the time and the third dimention are the nodes of the graph.}

\item{method}{a charcter string specifying the method}

\item{P}{a Pmat object from \code{permuco}}

\item{graph}{a igraph object of an undirected graph specifing the neighborgoods relationship between the nodes}

\item{coding_sum}{logical. If \code{TRUE}, set the coding of the design to sum, if \code{FALSE}, take the coding define in the dataframe.}

\item{multcomp}{the multiple comparison procedure only \code{"maris_oostenveld"} is available.}

\item{return_distribution}{If set to true return the null distribution by permutation.}

\item{...}{further arguments}
}
\description{
compute the maris oostenveld permutation test of on multiple signal. Used for full scalp EEG analysis.
}
