% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/image.clustergraph.R
\name{image.clustergraph}
\alias{image.clustergraph}
\title{Plot a statistical map of one effect.}
\usage{
\method{image}{clustergraph}(x, effect = 1, main = NULL, ylab = "", xlab = "", ...)
}
\arguments{
\item{x}{a clustergraph model.}

\item{effect}{an integer indicating which effect to plot.}

\item{main}{a character string indicating the name of the graphics. The default, NULL, get the name of the effect.}

\item{ylab}{see par. Default is "".}

\item{xlab}{see par. Default is "".}

\item{...}{other argument pass to image().,}
}
\description{
plot the significant test for 1 effect, all timepoints and all electrodes. Electrode are in the y-axis and the time in the x-axis. Non-significant cluster are shown in grey and the significant one in color from yellow to red as a function of the individual statistics.
}
