% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustergraph_rnd.R
\name{clustergraph_rnd}
\alias{clustergraph_rnd}
\title{Computer clustermass test based on multiple signal.}
\usage{
clustergraph_rnd(
  formula,
  data,
  signal,
  method,
  threshold,
  np,
  P,
  graph,
  effect = NULL,
  coding_sum = T,
  aggr_FUN = sum,
  multcomp = "maris_oostenveld",
  return_distribution,
  ...
)
}
\arguments{
\item{formula}{the formula of the model}

\item{data}{a dataframe containing the design}

\item{signal}{a 3 dimentional array. The the row are the observations, colomn the time and the third dimention are the nodes of the graph.}

\item{method}{a charcter string specifying the method}

\item{threshold}{see \code{clusterlm}.}

\item{np}{a scalar indicating the number of permutations. Will be overwrite by \code{P} if specified.}

\item{P}{a Pmat object from \code{permuco}}

\item{graph}{a igraph object of an undirected graph specifing the neighborgoods relationship between the nodes}

\item{effect}{a number indicating the effect to test. Refer to the \code{assign} attribute of the \code{model.matrix} object. If \code{NULL} it will test all the effects.}

\item{coding_sum}{logical. If \code{TRUE}, set the coding of the design to sum, if \code{FALSE}, take the coding define in the dataframe.}

\item{aggr_FUN}{the function to aggregate individual statistics into cluster mass.}

\item{multcomp}{the multiple comparison procedure only \code{"maris_oostenveld"} is available.}

\item{return_distribution}{If set to true return the null distribution by permutation.}

\item{...}{further arguments}
}
\description{
compute the maris oostenveld permutation test of on multiple signal. Used for full scalp EEG analysis.
}
